requirements.txt
langchain
langchain_community
langchain_experimental
groq
streamlit
pandas
numpy
langchain_groq
pillow
speechrecognition 
pyttsx3 
python-dotenv
apscheduler
transformers
googletrans==3.1.0a0
requests
matplotlib
faiss-cpu
sentence-transformers
pypdf
beautifulsoup4
duckduckgo-search
python-dateutil